fix interaction-requests urls, router
