+++ web/source/settings/views/user/router.tsx
@@ -26,8 +26,7 @@ import UserMigration from "./migration";
 import PostSettings from "./posts";
 import EmailPassword from "./emailpassword";
 import ExportImport from "./export-import";
-import InteractionRequests from "./interactions";
-import InteractionRequestDetail from "./interactions/detail";
+import InteractionRequestsRouter from "./interactions";
 
 /**
  * - /settings/user/profile
@@ -35,7 +34,7 @@ import InteractionRequestDetail from "./interactions/detail";
  * - /settings/user/emailpassword
  * - /settings/user/migration
  * - /settings/user/export-import
- * - /settings/users/interaction_requests
+ * - /settings/users/interaction-requests
  */
 export default function UserRouter() {
 	const baseUrl = useBaseUrl();
@@ -52,32 +51,10 @@ export default function UserRouter() {
 						<Route path="/emailpassword" component={EmailPassword} />
 						<Route path="/migration" component={UserMigration} />
 						<Route path="/export-import" component={ExportImport} />
-						<InteractionRequestsRouter />
-						<Route><Redirect to="/profile" /></Route>
+						<Route path="/interaction-requests/*?" component={InteractionRequestsRouter} />
+						<Redirect to="/profile" />
 					</Switch>
 				</ErrorBoundary>
-			</Router>
-		</BaseUrlContext.Provider>
-	);
-}
-
-/**
- * - /settings/users/interaction_requests/search
- * - /settings/users/interaction_requests/{reqId}
- */
-function InteractionRequestsRouter() {
-	const parentUrl = useBaseUrl();
-	const thisBase = "/interaction_requests";
-	const absBase = parentUrl + thisBase;
-
-	return (
-		<BaseUrlContext.Provider value={absBase}>
-			<Router base={thisBase}>
-				<Switch>
-					<Route path="/search" component={InteractionRequests} />
-					<Route path="/:reqId" component={InteractionRequestDetail} />
-					<Route><Redirect to="/search"/></Route>
-				</Switch>
 			</Router>
 		</BaseUrlContext.Provider>
 	);
